local/local_agent/requirements.txt
# Core Dependencies
langchain>=0.1.0
langgraph>=0.0.40
openai>=1.0.0
anthropic>=0.8.0

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Web Framework & API
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Database & Storage
sqlalchemy>=2.0.0
sqlite3
redis>=5.0.0
pydantic>=2.5.0

# Code Analysis & Processing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
ast-tools>=0.1.0
gitpython>=3.1.0

# Security & Sandboxing
docker>=6.1.0
psutil>=5.9.0
pathlib-mate>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
rich>=13.7.0
loguru>=0.7.0
asyncio-mqtt>=0.13.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: Local LLM Support
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Optional: Advanced Features
celery>=5.3.0
flower>=2.0.0
prometheus-client>=0.19.0
